Feat: 폴더 삭제를 위한 코드 작성 중

+++ components/modal/DeleteFolderModal.tsx
@@ -4,8 +4,7 @@ import ModalContainer from "./modalComponents/ModalContainer";
 import { deleteFolder } from "@/lib/api/folder";
 import toast from "react-hot-toast";
 import toastMessages from "@/lib/toastMessage";
-import { getLink } from "@/lib/api/link";
-import { useState } from "react";
+import useLinkListByFolderId from "@/hooks/useLinkListByFolderId";
 
 const DeleteFolderModal = ({
   folderName,
@@ -15,32 +14,21 @@ const DeleteFolderModal = ({
   folderId: number;
 }) => {
   const { closeModal } = useModalStore();
-  const [linkCount, setLinkCount] = useState(0);
-
-  // 폴더 내 링크 개수 확인 -> hook으로 빼기
-  const getLinkCountByFolder = async () => {
-    let query;
-    try {
-      const res = await getLink(query, folderId);
-      const data = res.data;
-      console.log(res);
-      console.log(data);
-      // setLinkCount()
-    } catch (err) {
-      console.error(err);
-    }
-  };
+  const { linkCount } = useLinkListByFolderId(folderId);
 
   const handleSubmit = async () => {
-    // 폴더 내에 링크 개수 0 일때만 삭제 가능 -> 링크 1개 이상이면 error toast 띄우겠습니다.
-    getLinkCountByFolder();
-    try {
-      await deleteFolder(folderId);
-      toast.success(toastMessages.success.deleteFolder);
-    } catch (error) {
-      toast.error(toastMessages.error.deleteFolder);
-    } finally {
-      closeModal();
+    // 폴더 내에 링크 개수 0 일때만 폴더 삭제 가능 -> 링크 1개 이상이면 error toast 띄우겠습니다. or 전체 링크 삭제 후 폴더 삭제
+    if (linkCount === 0) {
+      try {
+        await deleteFolder(folderId);
+        toast.success(toastMessages.success.deleteFolder);
+      } catch (error) {
+        toast.error(toastMessages.error.deleteFolder);
+      } finally {
+        closeModal();
+      }
+    } else {
+      toast.error(toastMessages.error.deleteNonEmptyFolder);
     }
   };
